+++ app/upload_step/upload_image_step.go
@@ -21,7 +21,7 @@ type UploadImageStepImpl struct {
 
 // FileInfo structure for output that matches Drupal's format
 type FileInfo struct {
-    FileID    int `json:"file_id"`
+    FileID    int64  `json:"file_id"`
     URI       string `json:"uri"`
     URL       string `json:"url"`
     MimeType  string `json:"mime_type"`
